chore: refactor coding keys interface
